// src/api/generated/model/orderResponseStatus.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * 뭐든배달 API 명세서
 * 뭐든배달 API 명세서입니다
 * OpenAPI spec version: v1
 */

export type OrderResponseStatus = (typeof OrderResponseStatus)[keyof typeof OrderResponseStatus];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const OrderResponseStatus = {
  CREATED: 'CREATED',
  PENDING: 'PENDING',
  PREPARING: 'PREPARING',
  RIDER_ASSIGNED: 'RIDER_ASSIGNED',
  DELIVERING: 'DELIVERING',
  COMPLETED: 'COMPLETED',
  REJECTED: 'REJECTED',
  CANCELED: 'CANCELED',
  CANCELLATION_REQUESTED: 'CANCELLATION_REQUESTED',
  PAYMENT_FAILED: 'PAYMENT_FAILED',
} as const;